Reorganize and remove wrapper
